some small updates

+++ week-08/day-4/Reddit/src/main/java/com/greenfox/reddit/services/PostServiceImpl.java
@@ -62,15 +62,15 @@ public class PostServiceImpl implements PostService{
     public int validPage(int nextPage) {
         if (nextPage <= 0) {
             return 0;
-        } else if (nextPage > (int) ((postRepository.findAll().size() -1) / 10 )) {
-            return (int) ((postRepository.findAll().size() -1) / 10);
+        } else if (nextPage > ((postRepository.findAll().size() -1) / 10 )) {
+            return (postRepository.findAll().size() -1) / 10;
         }
         return nextPage;
     }
 
     @Override
     public int[] maxPage() {
-        int highestPage = (int) ((postRepository.findAll().size() -1) / 10 ) + 1;
+        int highestPage = ((postRepository.findAll().size() -1) / 10 ) + 1;
         return java.util.stream.IntStream.range(1, highestPage + 1).toArray();
     }
